Listen for GETs on a per-guild basis
Pending GETs are now recognized regardless of channel. The `channels`
configuration key now controls which channels that GETs may be received
in.

When a build notification is sent, a flag is set for the entire guild,
meaning that the GET for that notification may be claimed in any channel
in the guild.